Refactor HapiBinaryAdapter to extend HapiAdapter
